Refactor ICMP ping
