+++ CMakeLists.txt
@@ -230,8 +230,9 @@ if (BUILD_TESTS AND NOT SKIP_SRC)
   find_package (cgreen REQUIRED)
 
   add_custom_target (tests
-    DEPENDS array-test alivedetection-test boreas_error-test networking-test 
-            xmlutils-test version-test osp-test nvti-test hosts-test)
+    DEPENDS array-test alivedetection-test boreas_error-test ping-test 
+            networking-test xmlutils-test version-test osp-test nvti-test 
+            hosts-test)
 
 endif (BUILD_TESTS AND NOT SKIP_SRC)
